Swal

+++ main.js
@@ -21,6 +21,11 @@ function saveIssue(e) {
     CreateDate: issueDate,
 
   }
+  Swal.fire(
+    'Good job!',
+    'You clicked the button!',
+    'success'
+  )
   // console.log(issue)
 
   if (localStorage.getItem('issues') == null) {
@@ -32,8 +37,6 @@ function saveIssue(e) {
     issues.push(issue);
     localStorage.setItem('issues', JSON.stringify(issues));
   }
-
-  document.getElementById('issueInputForm').reset();
 
   fetchIssues();
 
@@ -60,6 +63,11 @@ function setStatusClosed(id) {
       issues[i].status = 'Closed';
     }
   }
+  Swal.fire(
+    'Good job!',
+    'You Closed the status!',
+    'success'
+  )
 
   localStorage.setItem('issues', JSON.stringify(issues));
 
@@ -69,22 +77,58 @@ function setStatusClosed(id) {
 function deleteIssue(id) {
   var issues = JSON.parse(localStorage.getItem('issues'));
 
-  for (var i = 0; i < issues.length; i++) {
-    if (issues[i].id == id) {
-      issues.splice(i, 1);
+  Swal.fire({
+    title: 'Are you sure?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      Swal.fire(
+        'Deleted!',
+        'Your file has been deleted.',
+        'success'
+      )
+      for (var i = 0; i < issues.length; i++) {
+        if (issues[i].id == id) {
+          issues.splice(i, 1);
+        }
+      }
     }
-  }
+    localStorage.setItem('issues', JSON.stringify(issues));
 
-  localStorage.setItem('issues', JSON.stringify(issues));
-
-  fetchIssues();
+    fetchIssues();
+  })
 }
+
+// function OKSwal() {
+//   if(doc) {
+//     Swal.fire({
+//       title: 'Custom width, padding, background.',
+//       width: 600,
+//       padding: '3em',
+//       background: '#fff url(/images/trees.png)',
+//       backdrop: `
+//         rgba(0,0,123,0.4)
+//         url("./Image/tenor-cat.gif")
+//         left top
+//         no-repeat
+//       `
+//     })
+//   }
+//   }
+
 
 function fetchIssues() {
   var issues = JSON.parse(localStorage.getItem('issues'));
-  // var issuesListe = document.getElementById('issuesList');
 
-  issuesList.innerHTML = '';
+  ((issues == '')? issuesList.innerHTML = '<h4 style="margin: 12em 5em 30% 45%; color: gray;">No data.</h4>' : issuesList.innerHTML = '')
+ 
+
+  
 
   for (var i = 0; i < issues.length; i++) {
     var id = issues[i].id;
@@ -102,7 +146,8 @@ function fetchIssues() {
 
     issuesList.innerHTML +=   '<div class="well">'+
                               '<h6>Issue ID: ' + id + '</h6>'+
-                              '<p><span class="label label-info">' + status + '</span></p>'+
+                              // '<p><span class="label label-info">' + status + '</span></p>'+
+                              ((status == 'Open')? '<p><span class="label label-info">' + status + '</span></p>': '<p><span class="label label-default">' + status + '</span></p>') +
                               '<h3>' + desc + '</h3>'+
                               '<p><span class="glyphicon glyphicon-time"></span> ' + severity + '  ,' + CreateDate + '</p>'+
                               '<p><span class="glyphicon glyphicon-user"></span> ' + assignedTo + '</p>'+
